Move activity component
