+++ app/components/scoreboards/scoreboard_activity_component/scoreboard_activity_component.scss
