+++ spec/components/scoreboards/scoreboard_activity_component_spec.rb
@@ -2,38 +2,29 @@
 
 require 'rails_helper'
 
-RSpec.describe ScoreboardActivityComponent, :include_url_helpers, type: :component do
-  subject(:component) { described_class.new(**params) }
+RSpec.describe Scoreboards::ScoreboardActivityComponent, :include_url_helpers, type: :component do
+  subject(:html) do
+    render_inline(described_class.new(**params))
+  end
 
-  let(:id) { 'custom-id' }
-  let(:classes) { 'extra-classes' }
-  let(:show_positions) { true }
-  let!(:scoreboard) { create :scoreboard }
-  let!(:school) { create :school, scoreboard: scoreboard }
-  let!(:other_school) { create :school, scoreboard: scoreboard }
+  let!(:other_school) { create :school, scoreboard: create(:scoreboard) }
   let!(:activity_type) { create :activity_type, score: 10 }
   let!(:activity) { create(:activity, school: other_school, activity_type:) }
   let!(:observation) { create(:observation, :activity, activity: activity, school: other_school) }
 
-  let(:podium) { Podium.create(school: school, scoreboard: scoreboard) }
-
   let(:params) do
     {
       observations: [observation],
-      podium: podium,
-      id: id,
-      classes: classes,
-      show_positions: show_positions
+      podium: Podium.create(school: create(:school), scoreboard: other_school.scoreboard),
+      id: 'custom-id',
+      classes: 'extra-classes',
+      show_positions: true
     }
   end
 
-  let(:html) do
-    render_inline(component)
-  end
-
   it_behaves_like 'an application component' do
-    let(:expected_classes) { classes }
-    let(:expected_id) { id }
+    let(:expected_classes) { 'extra-classes' }
+    let(:expected_id) { 'custom-id' }
   end
 
   it 'displays the observations' do
@@ -48,7 +39,15 @@ RSpec.describe ScoreboardActivityComponent, :include_url_helpers, type: :compone
   it { expect(html).to have_link(activity.display_name, href: school_activity_path(other_school, activity)) }
 
   context 'when not showing positions' do
-    let(:show_positions) { false }
+    let(:params) do
+      {
+        observations: [observation],
+        podium: Podium.create(school: create(:school), scoreboard: other_school.scoreboard),
+        id: 'custom-id',
+        classes: 'extra-classes',
+        show_positions: false
+      }
+    end
 
     it 'displays the observations' do
       expect(html).not_to have_content(I18n.t('common.labels.place'))
